+++ CHANGES.rst
@@ -4,7 +4,8 @@ Products.MeetingCharleroi Changelog
 4.1.8 (unreleased)
 ------------------
 
-- Nothing changed yet.
+- Added a default classifier when an item is sent from one MeetingConfig to another.
+  [aduchene]
 
 
 4.1.7 (2020-10-30)